refactor/image: 이미지 확장자 webp로 변경 최적화 (#18)
* remove: png 이미지 삭제

* refactor: 로고 이미지 webp 확장자로 변경, 적용

* refactor: 구글아이콘 이미지 webp 확장자로 변경, 적용
