+++ public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="./sel-q-logo.png" />
+    <link rel="icon" href="./sel-q-logo.webp" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Sel-Q, 셀큐</title>
     <meta property="og:type" content="website" />
